requirements.txt
python-dateutil
coverage
nose
